move useActiveAnchors to util
